Maj change: Resets html + xsl-fo output to use the masters namespacing. This reverts the bc break that was added in the prvious version.
