+++ src/PaymentPart/Output/XslFoOutput/XslFoOutput.php
@@ -1,16 +1,17 @@
 <?php
 
-namespace Sprain\SwissQrBill\PaymentPart\Output\MarkupOutput;
+namespace Sprain\SwissQrBill\PaymentPart\Output\XslFoOutput;
 
-use Sprain\SwissQrBill\PaymentPart\Output\MarkupOutput\Template\Html\NewlineElementTemplate;
-use Sprain\SwissQrBill\PaymentPart\Output\MarkupOutput\Template\Html\PlaceholderElementTemplate;
-use Sprain\SwissQrBill\PaymentPart\Output\MarkupOutput\Template\Html\PrintableStylesTemplate;
-use Sprain\SwissQrBill\PaymentPart\Output\MarkupOutput\Template\Html\TextElementTemplate;
-use Sprain\SwissQrBill\PaymentPart\Output\MarkupOutput\Template\Html\PaymentPartTemplate;
-use Sprain\SwissQrBill\PaymentPart\Output\MarkupOutput\Template\Html\TitleElementReceiptTemplate;
-use Sprain\SwissQrBill\PaymentPart\Output\MarkupOutput\Template\Html\TitleElementTemplate;
+use Sprain\SwissQrBill\PaymentPart\Output\AbstractMarkupOutput;
+use Sprain\SwissQrBill\PaymentPart\Output\XslFoOutput\Template\PlaceholderElementTemplate;
+use Sprain\SwissQrBill\PaymentPart\Output\XslFoOutput\Template\PrintableStylesTemplate;
+use Sprain\SwissQrBill\PaymentPart\Output\XslFoOutput\Template\TextElementTemplate;
+use Sprain\SwissQrBill\PaymentPart\Output\XslFoOutput\Template\PaymentPartTemplate;
+use Sprain\SwissQrBill\PaymentPart\Output\XslFoOutput\Template\TitleElementTemplate;
+use Sprain\SwissQrBill\PaymentPart\Output\XslFoOutput\Template\TitleElementReceiptTemplate;
+use Sprain\SwissQrBill\PaymentPart\Output\XslFoOutput\Template\NewlineElementTemplate;
 
-final class HtmlOutput extends AbstractMarkupOutput
+final class XslFoOutput extends AbstractMarkupOutput
 {
     /**
      * @return string
